Fix some issue, and optimised code
