+++ src/Controller/Payment/PaymentController.java
@@ -1,8 +1,11 @@
 package Controller.Payment;
 
 import BusinessLogic.Payment.PaymentManager;
+import Exceptions.Search.GetAllPaymentYearsException;
+import Exceptions.Search.SearchPaymentException;
 import Model.Payment.Payment;
 
+import java.sql.Date;
 import java.util.ArrayList;
 
 public class PaymentController
@@ -13,6 +16,16 @@ public class PaymentController
     {
         return paymentManager.getAllPayments();
     }
+
+    public static ArrayList<Integer> getAllPaymentYears() throws GetAllPaymentYearsException {
+        return paymentManager.getAllPaymentYears();
+    }
+
+    // Search for payments based on criteria (validated, amount, year)
+    public static ArrayList<Payment> searchPayments(String paymentStatus, double minAmount, Date year) throws SearchPaymentException {
+        return paymentManager.searchPayments(paymentStatus, minAmount, year);  // Call the Payments Manager
+    }
+
     public static Payment getPayment(int id)
     {
         return paymentManager.getPayment(id);
