+++ src/DataAccess/Payment/PaymentDataAccess.java
@@ -7,8 +7,11 @@ import java.util.ArrayList;
 public interface PaymentDataAccess
 {
     ArrayList<Payment> getAllPayments();
-    Payment getPayment(int id);
+    ArrayList<Integer> getAllPaymentYears() throws Exceptions.Search.GetAllPaymentYearsException;
+    ArrayList<Payment> searchPayment(String status, double minAmount, java.sql.Date year) throws Exceptions.Search.SearchPaymentException;
+
     void createPayment(Payment payment);
     void updatePayment(Payment payment);
     void deletePayment(int id);
+    Payment getPayment(int id);
 }
