+++ src/Exceptions/Payment/GetAllPaymentException.java
@@ -1,0 +1,8 @@
+package Exceptions.Payment;
+
+public class GetAllPaymentException extends RuntimeException {
+    public GetAllPaymentException()
+    {
+        super("Error while getting all payments");
+    }
+}
